pulse/requirements.txt
django==4.2.26
djangorestframework
django-filter
drf-spectacular

mysqlclient
pymongo
sqlparse

celery
django-celery-beat
django-celery-results
pika

requests
beautifulsoup4
cryptography
python-dotenv

gunicorn